Documents/iacounter/requirements.txt
numpy>=1.21.3
opencv-python>=4.5.4.60
scikit-learn>=1.0.2
scipy>=1.7.3
pillow>=8.3.2
matplotlib>=3.5.0
pandas>=1.3.4
flask>=2.0.1
werkzeug>=2.0.1 